Update configuration and refactor model classes for improved functionality and performance

+++ dataset.py
@@ -76,11 +76,11 @@ class BilingualDataset(Dataset):
         assert label.size(0) == self.seq_len
 
         return {
-            "encoder_input": encoder_input,  # (seq_len)
-            "decoder_input": decoder_input,  # (seq_len)
+            "encoder_input": encoder_input.long(),  # (seq_len)
+            "decoder_input": decoder_input.long(),  # (seq_len)
             "encoder_mask": (encoder_input != self.pad_token).unsqueeze(0).unsqueeze(0).int(), # (1, 1, seq_len)
             "decoder_mask": (decoder_input != self.pad_token).unsqueeze(0).int() & causal_mask(decoder_input.size(0)), # (1, seq_len) & (1, seq_len, seq_len),
-            "label": label,  # (seq_len)
+            "label": label.long(),  # (seq_len)
             "src_text": src_text,
             "tgt_text": tgt_text,
         }
